Add nh definition
